Fix rubocop issues and refactor multipart uploader
- Extract multipart upload helpers into separate module to reduce class length
- Refactor upload_chunks method to reduce ABC complexity
- Split complex methods into smaller, focused methods
- Fix all rubocop offenses (no violations remaining)
- Maintain all functionality and tests passing

All tests passing: 310 examples, 0 failures
Rubocop clean: 134 files inspected, no offenses detected
